+++ src/app/components/calendar/calendar.component.scss
@@ -14,40 +14,6 @@
   min-width: 600px;
 }
 
-.event-sidebar {
-  flex: 1.2;
-  min-width: 400px;
-  background-color: #f5f5f5;
-  padding: 20px;
-  border-radius: 8px;
-
-  h3 {
-    margin-bottom: 10px;
-  }
-
-  ul {
-    list-style: none;
-    padding: 0;
-
-    li {
-      padding: 8px;
-      margin-bottom: 5px;
-      cursor: pointer;
-      background-color: #ecf0f1;
-
-      &:hover {
-        background-color: #dcdde1;
-      }
-    }
-  }
-}
-.event-description {
-  margin-top: 15px;
-  background-color: #ffffff;
-  padding: 10px;
-  border-left: 4px solid #3498db;
-  border-radius: 4px;
-}
 .calendar-nav {
   display: flex;
   gap: 10px;
@@ -69,6 +35,7 @@
 
 .calendar-header {
   margin-bottom: 10px;
+
   h2 {
     font-size: 1.8rem;
     color: #2c3e50;
