+++ src/app/components/events-sidebar/sidebar.component.html
@@ -1,3 +1,4 @@
+<H2>Upcoming Events</H2>
 <ul class="event-list">
   @for (event of events; track event.id) {
     <li
