+++ src/app/components/events-sidebar/sidebar.component.scss
@@ -3,6 +3,8 @@
   padding: 0;
   display: flex;
   flex-direction: column;
+  overflow-y: auto;
+  max-height: 80vh;
   gap: 12px;
 }
