+++ src/app/components/events-sidebar/sidebar.component.spec.ts
