+++ src/app/components/events-sidebar/sidebar.component.ts
